Front-End Web Developer Nanodegree Program - Build a Portfolio Site
Project 1 - Building a portfolio website. Replicate a design mockup
provided as a PDF in HTML and CSS. Develop a responsive website that
will display images, descriptions and links to each of the portfolio
projects completed through the course of the Udacity Front-End Web
Developer Nanodegree program on any size of screen.